requirements.txt

flask
pytesseract
Pillow
matplotlib
